pyproject.toml
[tool.poetry]
name = "python-postgressql"
version = "0.1.0"
description = ""
authors = ["Jonathan Chow <jonathan.chow@mclaren.com>"]

[tool.poetry.dependencies]
python = "^3.7"
psycopg2 = "^2.8.6"
pyyaml = "^5.4.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
